move methods to get mean and median metric from ROI to roi_utils
